move increaseTime function to tests
