+++ src/lib/eth.js
@@ -165,31 +165,6 @@ const Eth = {
       })
   },
 
-  increaseTimeTestRPC: function (seconds) {
-    return new Promise((resolve, reject) => {
-   	  function callback1(err, result){
-     		web3.currentProvider.send({
-               jsonrpc: '2.0',
-               method: 'evm_mine',
-                   params: [],
-                   id: new Date().getSeconds()
-               }, callback2);
-     	  };
-       function callback2(err, result){
-         if (err)
-           reject(err);
-         else
-           resolve(result);
-       };
-       web3.currentProvider.send({
-         jsonrpc: '2.0',
-         method: 'evm_increaseTime',
-         params: [seconds],
-         id: new Date().getSeconds()
-       }, callback1);
-     });
-   },
-
   getCurrentTimestamp: function() {
     return web3.eth.getBlockNumber()
       .then((blockNumber) => web3.eth.getBlock(blockNumber))
